+++ packages/terra-abstract-modal/CHANGELOG.md
@@ -4,7 +4,7 @@ Changelog
 Unreleased
 ----------
 ### Breaking Changes
-* Removed focus-trap form abstract modal
+* Removed focus-trap from abstract modal
 * Replaced with functionality that will disable focus on elements outside of the modal (not including content rendered from the modal in a portal)
 * Removed `isFocused` prop
 * Removed `fallbackFocus` prop. Focus now always shifts to modal container when opened.